+++ test/backgroundprocessor/BackgroundBlurProcessor.test.ts
@@ -919,16 +919,36 @@ describe('BackgroundBlurProcessor', () => {
     });
 
     it('can destroy', async () => {
-      backgroundFilterCommon.stubInit({ initPayload: 2, loadModelPayload: 2 });
+      const worker = sandbox.spy(
+        backgroundFilterCommon.stubInit({ initPayload: 2, loadModelPayload: 2 }),
+        'postMessage'
+      );
       stubSupported(true, false);
 
+      // Verify `stop` is called for BackgroundBlurProcessorBuiltIn
+      // when `blurCanvas` is null.
       let bbprocessor = (await BackgroundBlurVideoFrameProcessor.create()) as BackgroundBlurProcessorBuiltIn;
       bbprocessor['blurCanvas'] = null;
       await bbprocessor.destroy();
 
+      // Workers loaded for background blur must maintain API contract
+      // of `stop` command existing. Integration tests should exist on
+      // the worker code to ensure that it does not change. As a result,
+      // a unit test is fine here to ensure that `stop` contract is
+      // fulfilled on the client side.
+      expect(worker.callCount).to.be.equal(4);
+      expect(worker.calledWith({ msg: 'stop' })).to.be.true;
+      worker.resetHistory();
+
+      // Verify `stop` is called for BackgroundBlurProcessorBuiltIn
+      // when `blurCanvas` is not null.
       bbprocessor = (await BackgroundBlurVideoFrameProcessor.create()) as BackgroundBlurProcessorBuiltIn;
       bbprocessor['blurCanvas'] = document.createElement('canvas');
       await bbprocessor.destroy();
+
+      expect(worker.callCount).to.be.equal(4);
+      expect(worker.calledWith({ msg: 'stop' })).to.be.true;
+      worker.restore();
     });
 
     it('initialize handles errors', async () => {